requirements.txt
requests
pystyle
google-generativeai
beautifulsoup4
html5lib
urllib3
